+++ src/core.c
@@ -114,7 +114,7 @@ SEXP nano_encode(SEXP object) {
 
 }
 
-SEXP nano_encodes(SEXP data, SEXP mode, SEXP refhook) {
+SEXP nano_encodes(SEXP data, SEXP mode) {
 
   int xc;
   if (TYPEOF(mode) == INTSXP) {
@@ -142,19 +142,19 @@ SEXP nano_encodes(SEXP data, SEXP mode, SEXP refhook) {
   if (xc != 1)
     return nano_encode(data);
 
-  SEXP args, out;
+  SEXP out;
   switch (TYPEOF(data)) {
   case SYMSXP:
   case LANGSXP:
-    PROTECT(args = Rf_lang6(Rf_lang2(R_QuoteSymbol, data), R_NilValue, R_MissingArg, R_MissingArg, R_MissingArg, refhook));
-    PROTECT(out = Rf_lcons(nano_SerialSymbol, args));
+    PROTECT(out = Rf_lcons(nano_SerialSymbol,
+                           Rf_lang6(Rf_lang2(R_QuoteSymbol, data), R_NilValue, R_MissingArg, R_MissingArg, R_MissingArg, CAR(nano_refhook))));
     break;
   default:
-    PROTECT(args = Rf_lang6(data, R_NilValue, R_MissingArg, R_MissingArg, R_MissingArg, refhook));
-    PROTECT(out = Rf_lcons(nano_SerialSymbol, args));
+    PROTECT(out = Rf_lcons(nano_SerialSymbol,
+                           Rf_lang6(data, R_NilValue, R_MissingArg, R_MissingArg, R_MissingArg, CAR(nano_refhook))));
   }
   out = Rf_eval(out, R_BaseEnv);
-  UNPROTECT(2);
+  UNPROTECT(1);
 
   return out;
 
@@ -265,7 +265,7 @@ int nano_matchargs(SEXP mode) {
 
 }
 
-SEXP nano_decode(unsigned char *buf, size_t sz, const int mod, const int kpr, SEXP refhook) {
+SEXP nano_decode(unsigned char *buf, size_t sz, const int mod, const int kpr) {
 
   SEXP raw, data;
 
@@ -273,7 +273,7 @@ SEXP nano_decode(unsigned char *buf, size_t sz, const int mod, const int kpr, SE
     int tryErr = 0;
     PROTECT(raw = Rf_allocVector(RAWSXP, sz));
     memcpy(RAW(raw), buf, sz);
-    PROTECT(data = Rf_lang3(nano_UnserSymbol, raw, refhook));
+    PROTECT(data = Rf_lang3(nano_UnserSymbol, raw, CAR(nano_refhook)));
     data = R_tryEval(data, R_BaseEnv, &tryErr);
     if (tryErr) {
       data = raw;
@@ -658,7 +658,7 @@ SEXP rnng_listener_close(SEXP listener) {
 
 // send and recv ---------------------------------------------------------------
 
-SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
+SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block) {
 
   SEXP enc;
   R_xlen_t xlen;
@@ -672,7 +672,7 @@ SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
 
     if (block == R_NilValue) {
 
-      enc = nano_encodes(data, mode, refhook);
+      enc = nano_encodes(data, mode);
       xlen = Rf_xlength(enc);
       dp = RAW(enc);
       xc = nng_send(*sock, dp, xlen, NNG_FLAG_NONBLOCK);
@@ -680,7 +680,7 @@ SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
     } else if (TYPEOF(block) == LGLSXP) {
 
       const int blk = LOGICAL(block)[0];
-      enc = nano_encodes(data, mode, refhook);
+      enc = nano_encodes(data, mode);
       xlen = Rf_xlength(enc);
       dp = RAW(enc);
       xc = blk ? nng_send(*sock, dp, xlen, 0) : nng_send(*sock, dp, xlen, NNG_FLAG_NONBLOCK);
@@ -690,7 +690,7 @@ SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
       nng_msg *msgp;
       nng_aio *aiop;
       nng_duration dur = (nng_duration) Rf_asInteger(block);
-      enc = nano_encodes(data, mode, refhook);
+      enc = nano_encodes(data, mode);
       xlen = Rf_xlength(enc);
       dp = RAW(enc);
 
@@ -726,7 +726,7 @@ SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
     } else {
       dur = (nng_duration) Rf_asInteger(block);
     }
-    enc = nano_encodes(data, mode, refhook);
+    enc = nano_encodes(data, mode);
     xlen = Rf_xlength(enc);
     dp = RAW(enc);
 
@@ -793,7 +793,7 @@ SEXP rnng_send(SEXP con, SEXP data, SEXP mode, SEXP block, SEXP refhook) {
 
 }
 
-SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refhook) {
+SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes) {
 
   int xc;
   const int kpr = LOGICAL(keep)[0];
@@ -812,7 +812,7 @@ SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refh
       xc = nng_recv(*sock, &buf, &sz, NNG_FLAG_ALLOC + NNG_FLAG_NONBLOCK);
       if (xc)
         return kpr ? mk_error_recv(xc) : mk_error(xc);
-      res = nano_decode(buf, sz, mod, kpr, refhook);
+      res = nano_decode(buf, sz, mod, kpr);
       nng_free(buf, sz);
 
     } else if (TYPEOF(block) == LGLSXP) {
@@ -821,7 +821,7 @@ SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refh
       xc = blk ? nng_recv(*sock, &buf, &sz, NNG_FLAG_ALLOC): nng_recv(*sock, &buf, &sz, NNG_FLAG_ALLOC + NNG_FLAG_NONBLOCK);
       if (xc)
         return kpr ? mk_error_recv(xc) : mk_error(xc);
-      res = nano_decode(buf, sz, mod, kpr, refhook);
+      res = nano_decode(buf, sz, mod, kpr);
       nng_free(buf, sz);
 
     } else {
@@ -841,7 +841,7 @@ SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refh
       nng_aio_free(aiop);
       buf = nng_msg_body(msgp);
       sz = nng_msg_len(msgp);
-      res = nano_decode(buf, sz, mod, kpr, refhook);
+      res = nano_decode(buf, sz, mod, kpr);
       nng_msg_free(msgp);
     }
 
@@ -875,7 +875,7 @@ SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refh
     nng_aio_free(aiop);
     buf = nng_msg_body(msgp);
     sz = nng_msg_len(msgp);
-    res = nano_decode(buf, sz, mod, kpr, refhook);
+    res = nano_decode(buf, sz, mod, kpr);
     nng_msg_free(msgp);
 
   } else if (ptrtag == nano_StreamSymbol) {
@@ -922,7 +922,7 @@ SEXP rnng_recv(SEXP con, SEXP mode, SEXP block, SEXP keep, SEXP bytes, SEXP refh
 
     sz = nng_aio_count(aiop);
     nng_aio_free(aiop);
-    res = nano_decode(buf, sz, mod, kpr, refhook);
+    res = nano_decode(buf, sz, mod, kpr);
     R_Free(buf);
 
   } else {